#54 add unit tests for oM files
