+++ Test_Test/Test_Test.csproj
@@ -104,8 +104,17 @@
   <ItemGroup>
     <Compile Include="AdapterTests.cs" />
     <Compile Include="EngineTests.cs" />
-    <Compile Include="oMTests.cs" />
+    <Compile Include="oMTests\ContainsCopyright.cs" />
+    <Compile Include="oMTests\GetFiles.cs" />
+    <Compile Include="oMTests\IsPublicClass.cs" />
+    <Compile Include="oMTests\NameMatchesFileName.cs" />
+    <Compile Include="oMTests\NamespaceMatchesFilePath.cs" />
+    <Compile Include="oMTests\ObjectOnlyHasProperties.cs" />
+    <Compile Include="oMTests\PropertyAccessorsHaveNoBody.cs" />
+    <Compile Include="oMTests\PropertyHasPublicGet.cs" />
+    <Compile Include="oMTests\PropertyIsPublic.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="RunTest.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
@@ -114,6 +123,7 @@
     <Analyzer Include="..\packages\Microsoft.CodeAnalysis.Analyzers.2.9.4\analyzers\dotnet\cs\Microsoft.CodeAnalysis.Analyzers.dll" />
     <Analyzer Include="..\packages\Microsoft.CodeAnalysis.Analyzers.2.9.4\analyzers\dotnet\cs\Microsoft.CodeAnalysis.CSharp.Analyzers.dll" />
   </ItemGroup>
+  <ItemGroup />
   <Choose>
     <When Condition="'$(VisualStudioVersion)' == '10.0' And '$(IsCodedUITest)' == 'True'">
       <ItemGroup>
