+++ Test_Test/oMTests/NameMatchesFileName.cs
@@ -1,0 +1,20 @@
+﻿using System;
+using System;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace BH.Test.Test
+{
+    public partial class Test_oM
+    {
+        [TestMethod]
+        public void NameMatchesFileName()
+        {
+            Test.RunTest("NameMatchesFileName", GetChangedObjectFiles());
+        }
+    }
+}
